Please stay working now

+++ ipfish/src/analysis.ts
@@ -16,24 +16,30 @@ export interface AnalaysisBoard {
 }
 
 export class Analysis {
-  private static worker : Worker
+  private static worker : Worker | undefined
    
   public static async get(ns: NS, analysisBoard : AnalaysisBoard) : Promise<AnalysisState> {
     if (this.worker === undefined) {
-      this.worker = new analysisWorker()
       ns.atExit(() => {
-        this.worker.terminate()
+        if(this.worker !== undefined) this.worker.terminate()
+        this.worker = undefined
       }, "Analysis")
 
+      this.worker = new analysisWorker()
+
       const initalized = await new Promise((resolve, reject) => {
-        this.worker.onmessage = (event : MessageEvent<string>) => {
-          resolve(event.data) 
-        }
-        this.worker.onerror = (event) => {
-          reject(`Worker init onerror triggered ${event.message}`)
-        }
-        this.worker.onmessageerror = (event) => {
-          reject(`Worker init onmessageerror triggered ${event.data}`)
+        if (this.worker === undefined) {
+          reject("Worker non existent during init")
+        } else {
+          this.worker.onmessage = (event : MessageEvent<string>) => {
+            resolve(event.data) 
+          }
+          this.worker.onerror = (event) => {
+            reject(`Worker init onerror triggered ${event.message}`)
+          }
+          this.worker.onmessageerror = (event) => {
+            reject(`Worker init onmessageerror triggered ${event.data}`)
+          }
         }
       })
       ns.tprint(`Go worker ${initalized}`)
@@ -42,14 +48,18 @@ export class Analysis {
     this.worker.postMessage(analysisBoard)
 
     return new Promise((resolve, reject) => {
-      this.worker.onmessage = (event : MessageEvent<AnalysisState>) => {
-        resolve(event.data) 
-      }
-      this.worker.onerror = (event) => {
-        reject(`Worker onerror triggered ${event.message}`)
-      }
-      this.worker.onmessageerror = (event) => {
-        reject(`Worker onmessageerror triggered ${event.data}`)
+      if (this.worker === undefined) {
+        reject("Worker non existent during analysis")
+      } else {
+        this.worker.onmessage = (event : MessageEvent<AnalysisState>) => {
+          resolve(event.data) 
+        }
+        this.worker.onerror = (event) => {
+          reject(`Worker onerror triggered ${event.message}`)
+        }
+        this.worker.onmessageerror = (event) => {
+          reject(`Worker onmessageerror triggered ${event.data}`)
+        }
       }
     })
   }
